# Optimize readable groups query