This query is not very expensive in absolute terms but it done very often.

We'll always query for the world readable groups as a separate query, this only yields a few hundred groups.

For the case when we use a user.id join on the membership table instead of doing a subquery. We don't need to join on the user table as the previous version, just the membership one.

As a last step avoid going over the results to pluck the value, return them as a list of scalars directly from SQLA.

## Testing

TL;DR

For the old version:

```
Planning Time: 0.250 ms
￼
Execution Time: 72.769 ms
```

vs the new one


```
Planning Time: 0.295 ms
￼
Execution Time: 0.895 ms
```


```
-- Compare the old query plan
explain analyze SELECT "group".pubid AS group_pubid FROM "group" WHERE "group".readable_by = 'world' OR "group".readable_by = 'members' AND (EXISTS (SELECT 1 FROM user_group, "user" WHERE "group".id = user_group.group_id AND "user".id = user_group.user_id AND "user".id = 9264041))


-- To the new one
explain analyze SELECT anon_1.pubid 
FROM (SELECT "group".pubid AS pubid 
FROM "group" 
WHERE "group".readable_by = 'world' UNION SELECT "group".pubid AS pubid 
FROM "group" JOIN user_group ON "group".id = user_group.group_id 
WHERE user_group.user_id = 9264041 AND "group".readable_by = 'members') AS anon_1



-- Compare that they actually return the same values

SELECT "group".pubid AS group_pubid FROM "group" WHERE "group".readable_by = 'world' OR "group".readable_by = 'members' AND (EXISTS (SELECT 1 FROM user_group, "user" WHERE "group".id = user_group.group_id AND "user".id = user_group.user_id AND "user".id = 9264041))

except

 SELECT anon_1.pubid 
FROM (SELECT "group".pubid AS pubid 
FROM "group" 
WHERE "group".readable_by = 'world' UNION SELECT "group".pubid AS pubid 
FROM "group" JOIN user_group ON "group".id = user_group.group_id 
WHERE user_group.user_id = 9264041 AND "group".readable_by = 'members') AS anon_1

-- In both directions

SELECT anon_1.pubid 
FROM (SELECT "group".pubid AS pubid 
FROM "group" 
WHERE "group".readable_by = 'world' UNION SELECT "group".pubid AS pubid 
FROM "group" JOIN user_group ON "group".id = user_group.group_id 
WHERE user_group.user_id = 9264041 AND "group".readable_by = 'members') AS anon_1

except

SELECT "group".pubid AS group_pubid FROM "group" WHERE "group".readable_by = 'world' OR "group".readable_by = 'members' AND (EXISTS (SELECT 1 FROM user_group, "user" WHERE "group".id = user_group.group_id AND "user".id = user_group.user_id AND "user".id = 9264041))
```